cast the loss_mask to int32 in the `sparse_categorical_crossentropy` implementation
though the change made to Sum for bool is valid, we can't also update the Add operator to implicitly cast to int32 for bools, adding two bools is valid to result in a boolean in all ML libraries.

`TORCH=1 (or any interpreted backend) test/test_nn.py::TestNN::test_sparse_cat_cross_entropy`

+++ tinygrad/tensor.py
@@ -867,7 +867,7 @@ class Tensor:
 
   def sparse_categorical_crossentropy(self, Y:Tensor, ignore_index=-1) -> Tensor:
     # NOTE: self is a logits input
-    loss_mask = Y != ignore_index
+    loss_mask = (Y != ignore_index).cast(dtypes.int32)
     y_counter = Tensor.arange(self.shape[-1], dtype=dtypes.int32, requires_grad=False, device=self.device).unsqueeze(0).expand(Y.numel(), self.shape[-1])  # noqa: E501
     y = ((y_counter == Y.flatten().reshape(-1, 1)).where(-1.0, 0) * loss_mask.reshape(-1, 1)).reshape(*Y.shape, self.shape[-1])
     return self.log_softmax().mul(y).sum() / loss_mask.sum()
